final/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.5
pydantic-settings==2.10.1
PyJWT==2.10.1
pwdlib[argon2]==0.3.0
pyotp==2.9.0
qrcode[pil]==8.2
python-multipart==0.0.20
pandas==2.3.2
openpyxl==3.1.5
redis==6.4.0
celery==5.5.3
twilio==9.8.0
httpx==0.28.1
